이름 규칙 수정
