+++ src/main/java/com/catchbug/biz/account/AccountController.java
@@ -1,0 +1,5 @@
+package com.catchbug.biz.account;
+
+public class AccountController {
+
+}
