+++ src/main/java/com/catchbug/biz/board/BoardController.java
@@ -1,0 +1,5 @@
+package com.catchbug.biz.board;
+
+public class BoardController {
+
+}
