+++ src/main/java/com/catchbug/biz/vo/ProductVO.java
@@ -1,4 +1,4 @@
-package com.catchbug.biz.dto;
+package com.catchbug.biz.vo;
 
 import lombok.Data;
