+++ src/main/webapp/WEB-INF/views/admin/factory_franc_wait_list.jsp
