+++ src/main/webapp/WEB-INF/views/admin/factory_stock_list.jsp
