+++ src/main/webapp/WEB-INF/views/board/notice_board.jsp
